==> 03-numeric-exploration.Rmd <==
---
title: "03-numeric-exporation"
output: html_document
---
## Load the data
```{r}
library(data.table)
library(dplyr)
x = fread("data-sampled/train_numeric.csv") %>% as.data.frame()
response = fread("data-sampled/train_numeric.csv", select=c("Response"))$Response %>% factor()
id = x$Id
x$Id <- NULL
x$Response <- NULL
dim(x)
head(colnames(x))
tail(colnames(x))
```

## Inspection of the columns reveals them to be numerical with many NA values
Take a few columns at random
```{r}
summary(x[,2])
summary(x[, 100])
summary(x[, 700])
summary(x[, 960])
```

## Make a plot showing the number of NAs
```{r}
pctNA <- function(x){
  a = is.na(x)
  b = sum(a)
  c = length(x)
  d = (b/c)*100
  return(d)
}
percent_na = apply(x, 2, pctNA)
plot(percent_na)
```


## Plot the average of each feature
```{r}
y = apply(x, 2, mean, na.rm = T)
plot(y)
```

## Plot the standard deviation of each feature
```{r}
y = apply(x, 2, sd, na.rm = T)
plot(y)
```

## Boxplot each Line 0 feature

```{r}
library(tidyr)
x_long = gather(x, feature, value, L0_S0_F0:L3_S51_F4262)
df = filter(x_long, grepl("L0", feature))
boxplot(value ~ factor(feature), data = df)

```

Data look normalised

## Boxplot each Line 1 feature

```{r}
df = filter(x_long, grepl("L1", feature))
boxplot(value ~ factor(feature), data = df)

```

## Boxplot each Line 2 feature

```{r}
df = filter(x_long, grepl("L2", feature))
boxplot(value ~ factor(feature), data = df)

```

## Boxplot each Line 3 feature

```{r}
df = filter(x_long, grepl("L3", feature))
boxplot(value ~ factor(feature), data = df)

```
