==> 06-model-on-binary-vector.ipynb <==
import pandas as pd
train = pd.read_csv("data-transformed/05-for-bayes-sample.csv")

train.describe()

response = train.Response
del train['Id']
del train['Response']

response.describe()

### Niave Bayes Bernoulli

from sklearn.metrics import make_scorer
from sklearn.metrics import matthews_corrcoef
from sklearn import grid_search
from sklearn.naive_bayes import BernoulliNB

matt_score = make_scorer(matthews_corrcoef)

param_grid = {'alpha': [0, 0.25, 0.5, 0.75, 1]}

nb = BernoulliNB()

clf = grid_search.GridSearchCV(nb, param_grid, cv = 3, n_jobs = 3, scoring = matt_score)

clf.fit(train, response)
for params, mean_score, scores in clf.grid_scores_:
        print("%0.3f (+/-%0.03f) for %r"
              % (mean_score, scores.std() * 2, params))

### Linear SVM

from sklearn.svm import LinearSVC
matt_score = make_scorer(matthews_corrcoef)

param_grid = {'C': [0.1, 1, 10]
             }

svm = LinearSVC(dual = False, penalty = "l1", loss = 'squared_hinge', class_weight = 'auto') 
#Docs say make it dual false when n_samples > n_features
#Penality and loss have to be l1 and squared_hinge when using dual-False
#Also tried dual = True but no better

clf = grid_search.GridSearchCV(svm, param_grid, cv = 3, n_jobs = 3, scoring = matt_score)

clf.fit(train, response)
for params, mean_score, scores in clf.grid_scores_:
        print("%0.3f (+/-%0.03f) for %r"
              % (mean_score, scores.std() * 2, params))

### Random Forest

from sklearn.ensemble import RandomForestClassifier
matt_score = make_scorer(matthews_corrcoef)

param_grid = {'n_estimators': [500],
              'criterion': ['gini', 'entropy'],
              'min_samples_split': [2, 3, 4]
             }

rf = RandomForestClassifier() 
#Docs say make it dual false when n_samples > n_features
#Penality and loss have to be l1 and squared_hinge when using dual-False
#Also tried dual = True but no better

clf = grid_search.GridSearchCV(rf, param_grid, cv = 3, n_jobs = 3, scoring = matt_score)

clf.fit(train, response)
for params, mean_score, scores in clf.grid_scores_:
        print("%0.3f (+/-%0.03f) for %r"
              % (mean_score, scores.std() * 2, params))


Random Forest performs well. Just add these features to the main analysis vector

### RandomForest feature importance

from sklearn.ensemble import RandomForestClassifier

rf = RandomForestClassifier(n_estimators = 1000)
rf.fit(train, response)

importances = pd.DataFrame({'feature': train.columns.values,
                           'importances': rf.feature_importances_})
importances.sort('importances', ascending = False)
