==> 11-feature-importance-RF-sklearn.ipynb <==
import pandas as pd

df = pd.read_csv("data-transformed/sampled-data-joined.csv")

response = df.Response

del df['Response']
del df['Id']

from sklearn.ensemble import RandomForestClassifier

rf = RandomForestClassifier(n_estimators = 1000,
                           max_features = 100,
                           ).fit(df, response)

importances = rf.feature_importances_
colnames = df.columns.values

out = pd.DataFrame({"var": colnames,
                   "importance": importances})
out.to_csv("11-importances-output.csv", index = False)
